feat : 공백제거
